refactor: Move home_page to lib/ui directory
